+++ src/main.js
@@ -1,5 +1,5 @@
 'use strict';
 
 require('./form');
-require('./game');
+require('./game/game');
 require('./reviews/reviews');
